--- src/lib/i18n/messages/fr.json
@@ -1,165 +1,0 @@
-{
-	"nutriscore": "Nutri-Score",
-	"ecoscore": "Éco-Score",
-	"nova": "Nova - Niveaux de transformation",
-	"settings_link": "Paramètres",
-	"folksonomy_link": "Folksonomie",
-	"discover_link": "Découvrir",
-	"contribute_link": "Contribuer",
-	"producers_link": "Producteurs",
-	"prices_link": "Prix",
-	"login_link": "Connexion",
-	"logout_link": "Déconnexion",
-	"account_link": "Compte",
-	"world_option": "Monde",
-	"github_link": "GitHub",
-	"search": {
-		"title": "Recherche - {query} - Open Food Facts Explorer",
-		"placeholder": "Requête ou code-barres",
-		"go": "Aller",
-		"scan": "Scanner un code-barres",
-		"button": "Rechercher",
-		"product_not_found": "Aucun produit trouvé",
-		"product_not_found_desc": "Nous n'avons trouvé aucun produit correspondant à votre recherche",
-		"add_product": "Ajouter un nouveau produit",
-		"description": "Recherche sur Open Food Facts Explorer pour {query}"
-	},
-	"auth": {
-		"username": "Nom d'utilisateur",
-		"password": "Mot de passe",
-		"authenticated": "Authentifié",
-		"signin": "Se connecter",
-		"signout": "Se déconnecter",
-		"authenticating": "Authentification...",
-		"success": "Succès",
-		"failed": "Échec",
-		"enter_username": "Entrez le nom d'utilisateur",
-		"enter_password": "Entrez le mot de passe",
-		"status": "Statut",
-		"actions": "Actions"
-	},
-	"general": {
-		"language": "Langue",
-		"country": "Pays",
-		"currency": "Devise",
-		"loading": "Chargement...",
-		"warning": "Avertissement",
-		"notes": "Notes",
-		"summary": "Résumé",
-		"product_not_found": "Produit non trouvé",
-		"error_occurred": "erreur s'est produite :",
-		"errors_occurred": "erreurs se sont produites :"
-	},
-	"footer": {
-		"discover": {
-			"who_we_are": "Qui sommes-nous",
-			"blog": "Blog",
-			"code_of_conduct": "Code de conduite",
-			"vision_mission_values_and_programs": "Vision, mission, valeurs et programmes",
-			"partners": "Partenaires",
-			"faq": "FAQ",
-			"press": "Presse"
-		}
-	},
-	"settings": {
-		"general": "Général",
-		"influences": "Influences",
-		"login": "Connexion (sauvegardée dans localStorage) [NON SÉCURISÉ - DEBUG UNIQUEMENT]"
-	},
-	"folksonomy": {
-		"title": "Moteur de folksonomie",
-		"subtitle": "Explorateur de clés",
-		"description": "La folksonomie permet aux utilisateurs d'ajouter des étiquettes personnalisées aux produits. Parcourez toutes les clés disponibles ci-dessous, recherchez des clés spécifiques, ou cliquez sur une clé pour voir les produits qui l'utilisent.",
-		"search_placeholder": "Rechercher des clés...",
-		"keys_found": "clés trouvées",
-		"key_found": "clé trouvée",
-		"for": "pour"
-	},
-	"home": {
-		"welcome": "Bienvenue sur",
-		"welcome_user": "Bienvenue {username} sur",
-		"intro_1": "<strong>OpenFoodFacts Explorer</strong> est un outil pour explorer la base de données OpenFoodFacts.",
-		"intro_2": "Vous pouvez entrer un code produit dans la barre de recherche ci-dessus pour commencer ou cliquer sur l'un des produits ci-dessous."
-	},
-	"product": {
-		"title": "{productName} - Open Food Facts Explorer",
-		"description": "Page produit pour {productName} sur Open Food Facts Explorer",
-
-		"datasources": {
-			"title": "Sources de données",
-			"added_on": "Produit ajouté le {date} par : ",
-			"last_edit": "Dernière modification de la page produit le {date} par : ",
-			"also_edited_by": "Page produit également modifiée par :",
-			"last_check": "Dernière vérification de la page produit le {date} par : ",
-			"last_modified_t": "Dernière modification de la page produit le {date} par : ",
-			"incomplete_or_incorrect": "Si les données sont incomplètes ou incorrectes, vous pouvez les compléter ou les corriger en modifiant cette page.",
-			"unknown": "inconnu",
-			"done": "Terminé",
-			"toDo": "À faire"
-		},
-		"edit": {
-			"info": {
-				"images": "Téléchargez des images claires du produit, des ingrédients et des informations nutritionnelles pour aider les autres à identifier le produit.",
-				"basic_info": "Fournissez les principaux détails du produit, tels que la quantité, l'emballage et le fabricant.",
-				"languages": "Ajoutez les noms et descriptions du produit dans différentes langues pour le rendre accessible à plus d'utilisateurs.",
-				"ingredients": "Listez tous les ingrédients présents dans le produit. Cela aide les utilisateurs ayant des allergies ou des préférences alimentaires.",
-				"nutrition": "Saisissez les informations nutritionnelles telles qu'elles figurent sur l'emballage. Des données précises aident les utilisateurs à faire des choix éclairés.",
-				"comment": "Ajoutez des commentaires ou des notes concernant cette modification. Cela peut aider les réviseurs à comprendre vos changements."
-			},
-			"add_product": "Ajouter un produit",
-			"add_product_title": "Ajouter un nouveau produit",
-			"edit_product_title": "Modifier le produit",
-			"add_language": "Ajouter une langue",
-			"search_languages": "Rechercher des langues à ajouter",
-			"no_languages_found": "Aucune langue trouvée",
-			"name": "Nom",
-			"quantity": "Quantité",
-			"emb_code": "Code EMB (Code producteur)",
-			"packaging": "Emballage",
-			"manufacturing_places": "Lieux de fabrication",
-			"categories": "Catégories :",
-			"labels": "Labels",
-			"brands": "Marques",
-			"stores": "Magasins",
-			"origins": "Origines",
-			"countries": "Pays",
-			"website_url": "URL du site web",
-			"photos": "Photos du produit",
-			"no_photo_available": "Aucune photo disponible",
-			"ingredients": "Ingrédients",
-			"ingredients_list": "Liste des ingrédients",
-			"no_ingredients_image": "Aucune image d'ingrédients",
-			"nutritional_information": "Informations nutritionnelles",
-			"no_nutrition_image": "Aucune image nutritionnelle",
-			"serving_size": "Taille de portion",
-			"no_nutrition_data": "Aucune information nutritionnelle sur le produit",
-			"no_nutrition_specified": "Aucune donnée nutritionnelle spécifiée",
-			"comment": "Commentaire",
-			"comment_placeholder": "Ajouter un commentaire à cette modification",
-			"save_btn": "Enregistrer",
-			"debug": "Debug",
-			"sections": {
-				"images": "Images",
-				"basic_info": "Informations de base",
-				"languages": "Langues et noms du produit",
-				"ingredients": "Ingrédients",
-				"nutrition": "Informations nutritionnelles",
-				"comment": "Commentaire"
-			}
-		}
-	},
-	"errors": {
-		"network": {
-			"title": "Erreur de connexion",
-			"message": "Impossible de se connecter au serveur Open Food Facts. Veuillez vérifier votre connexion Internet et réessayer.",
-			"instructions": "Si le problème persiste, visitez la page de statut d'Open Food Facts."
-		}
-	},
-	"common": {
-		"next": "Suivant",
-		"back": "Retour",
-		"skip": "Ignorer",
-		"step": "Étape",
-		"of": "sur"
-	}
-}
